entidades agregadas

+++ src/application/use-cases/product.service.js
@@ -1,4 +1,5 @@
 const Product = require('../../domain/entities/product.entity');
+const { NotFoundError } = require('../../domain/errors');
 
 class ProductService {
     constructor(productRepository) { // ¡Depende de la interfaz!
@@ -10,7 +11,11 @@ class ProductService {
     }
 
     async getProductById(id) {
-        return this.productRepository.getById(id);
+        const product = await this.productRepository.getById(id);
+        if (!product) {
+            throw new NotFoundError(`Product with id ${id} not found`);
+        }
+        return product;
     }
 
     async createProduct(productData) {
@@ -22,12 +27,19 @@ class ProductService {
             productData.price,
             productData.stock,
             productData.category,
-            productData.imageUrl
+            productData.imageUrl,
+            //AQUI SE AGREGA EL CAMPO DEL EJERCICIO 1
+            productData.brand //agregando nuevo campo
         );
         return this.productRepository.create(productEntity);
     }
 
     async updateProduct(id, productData) {
+        const existingProduct = await this.productRepository.getById(id);
+        if (!existingProduct) {
+            throw new NotFoundError(`Product with id ${id} not found`);
+        }
+
         const productEntity = new Product(
             id,
             productData.name,
@@ -35,12 +47,18 @@ class ProductService {
             productData.price,
             productData.stock,
             productData.category,
-            productData.imageUrl
+            productData.imageUrl,
+            //AQUI SE AGREGA EL CAMPO DEL EJERCICIO 1
+            productData.brand //nuevo campo agregado
         );
         return this.productRepository.update(id, productEntity);
     }
 
     async deleteProduct(id) {
+        const product = await this.productRepository.getById(id);
+        if (!product) {
+            throw new NotFoundError(`Product with id ${id} not found`);
+        }
         return this.productRepository.delete(id);
     }
 }
